make left arrow and right arrow pass first and last index of each line to the next

+++ src/utilities.h
@@ -20,7 +20,7 @@
 
 struct string{
     char* b;
-    size_t len;
+    int len;
 };
 struct editorConfig {
     int cx,cy;
